requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.1
DateTime==4.3
h11==0.12.0
idna==3.3
numpy==1.21.2
outcome==1.1.0
pqdict==1.1.1
pycparser==2.21
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
zope.interface==5.4.0
